restructure sourcecode
